moved roles, enroll and registerrole into a seperate roleDistribution folder #2
