+++ src/main/java/chat/commands/roleDistribution/Roles.java
@@ -1,5 +1,6 @@
-package chat.commands;
+package chat.commands.roleDistribution;
 
+import chat.commands.Command;
 import sx.blah.discord.api.IDiscordClient;
 import sx.blah.discord.handle.obj.IMessage;
 import sx.blah.discord.handle.obj.Permissions;
@@ -7,8 +8,8 @@ import sx.blah.discord.util.MessageBuilder;
 
 import java.util.stream.Collectors;
 
-/**
- * Created by Dries on 3/07/2017.
+/*
+ * Dave replies with a list of his roles that he can distribute with the enroll command
  */
 public class Roles implements Command {
   /**
@@ -18,6 +19,7 @@ public class Roles implements Command {
   public void processCommand(IDiscordClient dave, IMessage message) {
     String rolesString = dave.getRoles().stream()
             .filter(specificRole -> !specificRole.getPermissions().contains(Permissions.ADMINISTRATOR))
+            .skip(1)
             .map(o -> o.getName().replace('@', ' '))
             .collect(Collectors.joining(", "));
     new MessageBuilder(dave)
